Rename 'components/StaticInfo.vue' to 'components/index/SiteTopUpper/StaticInfoCard.vue'
